+++ risk-service/src/main/java/com/acquirerx/risk/risk/dto/RiskRuleResponseDTO.java
@@ -2,13 +2,15 @@ package com.acquirerx.risk.risk.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class RiskRuleResponseDTO {
 
     private Long riskRuleId;
     private String name;
     private String expression;
-    private Double maxAmount;
+    private BigDecimal maxAmount;
     private String severity;
     private String action;
     private Boolean active;
