+++ settlement-service/src/main/java/com/acquirerx/settlement/settlement/dto/SettlementBatchResponseDTO.java
@@ -2,6 +2,7 @@ package com.acquirerx.settlement.settlement.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -10,9 +11,9 @@ public class SettlementBatchResponseDTO {
     private Long settleBatchId;
     private Long merchantId;
     private String merchantName;
-    private Double grossAmount;
-    private Double totalFees;
-    private Double netAmount;
+    private BigDecimal grossAmount;
+    private BigDecimal totalFees;
+    private BigDecimal netAmount;
     private Integer txnCount;
     private String status;
     private LocalDateTime periodStart;
