+++ settlement-service/src/main/java/com/acquirerx/settlement/settlement/dto/SettlementFilterDTO.java
@@ -2,6 +2,7 @@ package com.acquirerx.settlement.settlement.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -9,8 +10,8 @@ public class SettlementFilterDTO {
 
     private String status;
     private Long merchantId;
-    private Double minNetAmount;
-    private Double maxNetAmount;
+    private BigDecimal minNetAmount;
+    private BigDecimal maxNetAmount;
     private LocalDateTime fromDate;
     private LocalDateTime toDate;
     private Integer minTxnCount;
