+++ settlement-service/src/main/java/com/acquirerx/settlement/settlement/repository/AdjustmentRepository.java
@@ -7,10 +7,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
+
 public interface AdjustmentRepository extends JpaRepository<Adjustment, Long> {
 
     Page<Adjustment> findByMerchantId(Long merchantId, Pageable pageable);
 
     @Query("SELECT SUM(a.amount) FROM Adjustment a WHERE a.merchantId = :merchantId AND a.status = 'APPLIED'")
-    Double sumAdjustmentsByMerchant(@Param("merchantId") Long merchantId);
+    BigDecimal sumAdjustmentsByMerchant(@Param("merchantId") Long merchantId);
 }
