+++ transaction-service/src/main/java/com/acquirerx/transaction/fee/dto/TxnFeeResponseDTO.java
@@ -4,18 +4,19 @@ import com.acquirerx.transaction.switchmodule.enums.TxnStatus;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 @AllArgsConstructor
 public class TxnFeeResponseDTO {
     private Long txnId;
-    private Double amount;
-    private Double schemeFee;
-    private Double interchangeFee;
-    private Double acquirerMarkup;
-    private Double totalFee;
-    private Double netMerchantAmount;
+    private BigDecimal amount;
+    private BigDecimal schemeFee;
+    private BigDecimal interchangeFee;
+    private BigDecimal acquirerMarkup;
+    private BigDecimal totalFee;
+    private BigDecimal netMerchantAmount;
     private TxnStatus status;
     private LocalDateTime txnDate;
 }
