+++ transaction-service/src/main/java/com/acquirerx/transaction/fee/dto/TxnFilterDTO.java
@@ -2,6 +2,7 @@ package com.acquirerx.transaction.fee.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -9,8 +10,8 @@ public class TxnFilterDTO {
 
     private String status;
     private Boolean settled;
-    private Double minAmount;
-    private Double maxAmount;
+    private BigDecimal minAmount;
+    private BigDecimal maxAmount;
     private LocalDateTime fromDate;
     private LocalDateTime toDate;
     private Long merchantId;
