+++ transaction-service/src/main/java/com/acquirerx/transaction/fee/dto/TxnStatsDTO.java
@@ -3,12 +3,14 @@ package com.acquirerx.transaction.fee.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class TxnStatsDTO {
     private long totalTxns;
     private long settledTxns;
     private long unsettledTxns;
-    private double settledAmount;
-    private double totalFees;
+    private BigDecimal settledAmount;
+    private BigDecimal totalFees;
 }
