+++ transaction-service/src/main/java/com/acquirerx/transaction/switchmodule/dto/TransactionFilterDTO.java
@@ -2,6 +2,7 @@ package com.acquirerx.transaction.switchmodule.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -9,8 +10,8 @@ public class TransactionFilterDTO {
 
     private String status;
     private String txnType;
-    private Double minAmount;
-    private Double maxAmount;
+    private BigDecimal minAmount;
+    private BigDecimal maxAmount;
     private LocalDateTime fromDate;
     private LocalDateTime toDate;
     private Long merchantId;
